Group9_HW1/requirements.txt
matplotlib==3.8.0
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.5.0
Pillow==10.4.0
scipy==1.14.1
